Unicode fixes.
Do not use _(), this does not work as expected in UNICODE. RTFM! :)
Have a look at the proper way to translate static initialized strings.
And also notice the proper way to translate them in code with
wxGetTranslation().

Efe, maybe we should check the rest of the code to see if this wrong usage
happens elsewhere, unfortunately, I don't have the time to do that now :/ .