This commit have double function:
a) Fixed the crash on non-std locales because of date/time functions. This was happening because of a bug on
wxWidgets, so I report it and also make we use ISO data/time so we don't crash anymore.

b) Get rid of more ugly unicode2char stuff by making the logfile a wxFFileOutputStream. Also speedup on writing to
logfile.