Reworked the template finding routines of amuleweb.
Now amuleweb searches in for the given template in some directories:
 1. $HOME/.aMule/webserver
 On linux:
 	2. /usr/share/amule/webserver
	3. /usr/local/share/amule/webserver
MacOS guys and anybody are welcome to add their OS's default install paths
to the search path.

Template is now defined somewhat else:
the program looks for [SearchPath]/[TemplateName]/aMule.tmpl

So: template file name is always aMule.tmpl, and it's container directory
is the template name.

Default template name is 'default'.

Install is modified to reflect the new skin system.