Change the IPFilter class so that when an update is downloaded, it replaces the existing completly, instead of being merged with them as was currently done. This avoid the situation where spurious ranges are made as a result of merging two ipfilter.dat files.
In order to be able to accomondate ipfilter-ranges that a
user would want to keep between updates, a second ipfilter
file is now also used: ~/.aMule/ipfilter_static.dat
Unlike ~/.aMule/ipfilter.dat, this file is never changed
by aMule and is loaded _after_ the other ipfilter list, so
that it takes precedence over the normal ipfilter.dat file.

When the aMule is run now, a template file is created for
both ipfilter.dat and ipfilter_static.dat describing the
purpose of the file. This template is only created if the
files do not already exist.

Loading of zip-archives has also been changed. Instead of
reading the ipfiler-list directly from the archive, the
list is now extracted first and then parsed. This has 
resulted in overhead from reading zipped ipfilter-lists
compared to reading non-packed lists being vastly reduced.

Finally, the loading of the ipfilter files has been moved
so that it is done after the aMule GUI has been loaded.