Removing the Write* functions from CSafeMemFile, since they do nothing that the CMemFile class doens't already do by itself.
Also removed all but one of the Read* functions, since
those checks are already done in CMemFile. It should also 
be noted that the the read-checks currently checked against
the wrong size (the actual size of buffer, not the written
area), so we could potentially read uninitialized values in
cases where the file is both written to and read from and
the growth-value was non-zero. Luckily, I don't think we
have such usage ATM.

I left the checks in ReadUInt128, since they were needed
to ensure that the function is atomic.