Proper config location on Mac and Win.
In case an old configuration is detected, it will be moved silently to the new location,
unless the new location contains a valid configuration. When moving configuration,
config file will be edited to make sure every path in it is still valid after the move.

NOTE: config relocation is only done by amule(d)