Optimized and simplified adding of shared files to the listctrl:
On startup, both complete and part-files are now
added in one go and instead of inserting each in
its sorted position, which had O(n^2) time
complexity due to on average having to move half
the existing items when inserting, they are now
inserted last and then sorted afterwards.

On reloads after aMule has started, the files 
would be added twice. First indivially and then
removed and readded batch. This has been fixed
and now works as the startup loading.

This cut off a forth of the time used to load
shared files on my system.

I also dropped the setting of the text in the
cells of the sharedfilesctrl, since this
accounted for another fourth of the time used,
and the listctrl is ownerdrawn anyway, so there
was no gain by doing that.

These changes in total cuts the time used to
load shared files in half. The vast majority of
the remaining time is used to search for shared
files. This will be moved to a thread-task later.