Heavily modified IP2Country code.
C'mon people, let's keep some quality here.

1) Gui code on core files?
2) Duplicated code for remote gui and monolithic?
3) No threads and:
 3.1) Mutexes?
 3.2) Mutable vars
 3.3) C-like static creation of data?
4) Constant lookup of the name when we have a map for the flags?
5) Two functions used insteaf of returning a class, or at least a struct?
6) Etcetera.

Really. What the hell.

I also changed the displayed text to be the country code, not the name. The name was too long.
If someone feels like complaning about that change, go ahead and add another column for this. 

I am open to opinions about the code being shown as lowercase or uppercase. Further optimization
could include storing the country code in the client object, to avoid looking it up on every redraw.