Keep downloads and shares in a common list in the remotegui, and show completed downloads at last
This is a follow-up to 10056 (and fixes problems with downloads completing while remotegui is connected).
Completed downloads are now shown just like in aMule. Only problem is, they keep piling up if the core runs for some time. Whenever the gui reconnects, all completed downloads since core start are shown.