Fixed foolish crashbug introduced in 10194
It caused a throw that was somehow caught in the event handler, which then innocently continued processing the next EC request being in a wrong state and so breaking the EC connection.
Such things are fun to step through in the debugger, when you are suddenly somewhere and have no idea how you've got there. >:(

Have a good laugh at me...