Asio: configure support
Added support to configure to decide which socket library to use
(wxWidgets or Boost.Asio). The default is the wxWidgets socket library.
To use Boost.Asio configure aMule with --with-boost[=/path/to/boost].

If the boost sources are found, they'll be used. Otherwise aMule will be
linked to the Boost.System library.