Adapt configure to other VCSes
As nowadays it's way too easy to create a git or mercurial clone of a subversion
repository, we'd better be prepared for that. In case we find that we're building
from a working copy (checkout, working tree, whatever you prefer to name it), we
assume that the tools to handle it are available and are fairly up-to-date. (This
assumption already applied to subversion, it's now extended to git and mercurial.)

Format of SVNDATE:
1) If .svn-revision is present in the root of the source tree, then the contents of
   that file, "rev. " prepended.
2) In case subversion/git/mercurial is detected, then it's "rev. " plus
   2a) the subversion revision number, if it's a subversion working copy or the
       subversion revision can be derived from the git-svn/hgsubversion repository,
       plus
   2b) the git/mercurial revision id, if the subversion revision cannot be derived
       or there are local commits on top of the subversion revision, plus
   2c) a plus sign if the working copy is modified.
3) Otherwise (e.g. in case of release tarballs) it's empty.